aflux/requirements.txt
python-alipay-sdk
boto3
elasticsearch
amqp==2.5.0
aniso8601==7.0.0
asn1crypto==0.24.0
astroid==2.2.5
bcrypt==3.1.7
billiard==3.6.0.0
celery==4.3.0
cffi==1.12.3
Click==7.0
colorama==0.4.1
coverage==4.5.4
cryptography==2.8
entrypoints==0.3
flake8==3.7.8
flake8-polyfill==1.0.2
Flask==1.1.1
Flask-JWT-Extended==3.21.0
flask-mongoengine==0.9.5
flask-redis==0.4.0
Flask-RESTful==0.3.7
Flask-WTF==0.14.2
future==0.17.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.3
lazy-object-proxy==1.4.1
mando==0.6.4
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.18.2
mypy==0.720
mypy-extensions==0.4.1
nose==1.3.7
nose-parameterized==0.6.0
pika==1.1.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.8.0
pytz==2019.2
radon==3.0.3
redis==3.3.6
six==1.12.0
typed-ast==1.4.0
typing-extensions==3.7.4
vine==1.3.0
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
